Fully implement CRUD for employees

+++ Company/Company/EmployeesViewController.swift
@@ -1,14 +1,22 @@
 import UIKit
+import os.log
 
 class EmployeesViewController: UITableViewController {
     var employees: [Employee] = []
 
-    @IBAction func refreshEmployeesClicked(_ sender: Any) {
+    @IBAction func refreshEmployeesClicked(_ sender: Any?) {
         loadData()
+        tableView.reloadData()
+        refreshControl?.endRefreshing()
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        refreshControl = UIRefreshControl()
+        refreshControl!.addTarget(self, action: #selector(refreshEmployeesClicked), for: .valueChanged)
+        
+        navigationItem.leftBarButtonItem = editButtonItem
 
         // Uncomment the following line to preserve selection between presentations
         // self.clearsSelectionOnViewWillAppear = false
@@ -54,38 +62,55 @@ class EmployeesViewController: UITableViewController {
         
         let employee = employees[indexPath.row] as Employee
         cell.employee = employee
+        if employee.image != nil {
+            let url = URL(string: Api.companyImageUrl + employee.image!)
+            
+            Api.downloadImage(url: url!, completion: { (image: UIImage?, success: Bool, error: String?) in
+                if(success) {
+                    DispatchQueue.main.async {
+                        cell.employeeImage.image = image
+                    }
+                }
+                else {
+                    DispatchQueue.main.async {
+                        cell.employeeImage.image = image
+                    }
+                }
+                
+            })
+        } else {
+            print("Employee has no image")
+            cell.employeeImage.image = nil
+        }
         return cell
-    }
-    @IBAction func cancelToEmployeesViewController(_ segue: UIStoryboardSegue) {
-        
-    }
-    
-    @IBAction func saveEmployeeDetail(_ segue: UIStoryboardSegue) {
-        
     }
     
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        let employee = employees[indexPath.row] as Employee
         
-        Employee.deleteEmployee(id: employee.id) { (success) in
-            if(success) {
-                self.employees.remove(at: indexPath.row)
-                let indexPaths = [indexPath]
-                DispatchQueue.main.async(execute: {
-                    tableView.deleteRows(at: indexPaths, with: .automatic)
-                })
+        if editingStyle == .delete {
+            // Delete employee in api, update locally
+            let employee = employees[indexPath.row]
+            Employee.deleteEmployee(id: employee.id) { (success) in
+                if(success) {
+                    self.employees.remove(at: indexPath.row)
+                    let indexPaths = [indexPath]
+                    DispatchQueue.main.async(execute: {
+                        tableView.deleteRows(at: indexPaths, with: .fade)
+                    })
+                }
+                
             }
+        }
+        else if editingStyle == .insert {
             
-        }      
+        }
     }
 
-    /*
     // Override to support conditional editing of the table view.
     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
         // Return false if you do not want the specified item to be editable.
         return true
     }
-    */
 
     /*
     // Override to support editing the table view.
@@ -114,15 +139,78 @@ class EmployeesViewController: UITableViewController {
     }
     */
 
-    /*
+    
     // MARK: - Navigation
 
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
+        super.prepare(for: segue, sender: sender)
+        
+        switch(segue.identifier ?? "") {
+            
+        case "AddEmployee":
+            os_log("Adding a new employee", log: OSLog.default, type: .debug)
+            
+        case "ShowEmployeeDetail":
+            guard let employeeDetailsViewController = segue.destination as? EmployeeDetailsViewController else {
+                fatalError("Unexpected destination: \(segue.destination)")
+            }
+            
+            guard let selectedEmployeeCell = sender as? EmployeeCell else {
+                //fatalError("Unexpected sender: \(sender)")
+                fatalError("Unexpected sender: " + sender.debugDescription)
+            }
+            
+            guard let indexPath = tableView.indexPath(for: selectedEmployeeCell) else {
+                fatalError("The selected cell is not being displayed by the table")
+            }
+            
+            let selectedEmployee = employees[indexPath.row]
+            employeeDetailsViewController.employee = selectedEmployee
+            
+        default:
+            fatalError("Unexpected Segue Identifier")
+        }
     }
-    */
+    
+    // MARK: Actions
+    @IBAction func unwindToEmployeeList(sender: UIStoryboardSegue) {
+        if let sourceViewController = sender.source as? EmployeeDetailsViewController, let employee = sourceViewController.employee {
+            
+            if let selectedIndexPath = tableView.indexPathForSelectedRow {
+                // Update an existing employee
+                //employees[selectedIndexPath.row] = employee
+                //tableView.reloadRows(at: [selectedIndexPath], with: .none)
+                print("returning from editing")
+                print(employee)
+                // Send changes to api
+                Employee.updateEmployee(employee: employee) { (success) in
+                    if(success) {
+                        DispatchQueue.main.async(execute: {
+                            self.employees[selectedIndexPath.row] = employee
+                            self.tableView.reloadRows(at: [selectedIndexPath], with: .none)
+                        })
+                    }
+                }
+            }
+            else {
+                Employee.addEmployee(employee: employee) { (success) in
+                    if(success) {
+                        DispatchQueue.main.async(execute: {
+                            let newIndexPath = IndexPath(row: self.employees.count, section: 0)
+                            self.employees.append(employee)
+                            self.tableView.insertRows(at: [newIndexPath], with: .automatic)
+                        })
+                    }
+                    else {
+                        print("error in api")
+                    }
+
+                }
+            }
+
+        }
+    }
 
 }
